# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel and Swagger",
        "description": "base url image https://ubgmart.com/storage/ , s\u1eed d\u1ee5ng basic auth theo t\u00e0i kho\u1ea3n \u0111\u01b0\u1ee3c cung c\u1ea5p \u0111\u1ec3 truy c\u1eadp api",
        "termsOfService": "",
        "contact": {
            "email": "thanhan1507@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "api-ubg.test",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/list-order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "L\u1ea5y ra danh s\u00e1ch \u0111\u01a1n h\u00e0ng",
                "description": "Danh s\u00e1ch \u0111\u01a1n h\u00e0ng ",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID custommer login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "pending, canceled, completed",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/single-order": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Th\u00f4ng tin chi ti\u1ebft \u0111\u01a1n h\u00e0ng",
                "description": "Chi ti\u1ebft \u0111\u01a1n h\u00e0ng",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID \u0111\u01a1n h\u00e0ng",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Id customer Login",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/address-list": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Danh s\u00e1ch \u0111\u1ecba ch\u1ec9 c\u1ee7a user",
                "description": "Danh s\u00e1ch \u0111\u1ecba ch\u1ec9 c\u1ee7a user, hi\u1ec3n th\u1ecb trong l\u1ef1a ch\u1ecdn \u0111\u01a1n h\u00e0ng",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "query",
                        "description": "ID custommer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/get-coupon-code": {
            "get": {
                "tags": [
                    "Promotion & Discount"
                ],
                "summary": "L\u1ea5y ra th\u00f4ng tin gi\u1ea3m gi\u00e1 t\u1eeb m\u00e3 coupon",
                "description": "M\u00e3 coupon",
                "parameters": [
                    {
                        "name": "applied_coupon_code",
                        "in": "query",
                        "description": "M\u00e3 coupon",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/get-promotion-in-cart-value": {
            "get": {
                "tags": [
                    "Promotion & Discount"
                ],
                "summary": "L\u1ea5y ra s\u1ed1 ti\u1ec1n n\u1ebfu c\u00f3 ch\u01b0\u01a1ng tr\u00ecnh promotion",
                "description": "Promotion tr\u1eeb ti\u1ec1n trong \u0111\u01a1n h\u00e0ng, s\u1eed d\u1ee5ng \u0111\u1ec3 tr\u1eeb \u0111i t\u1ed5ng s\u1ed1 ti\u1ec1n \u0111\u01a1n h\u00e0ng",
                "parameters": [
                    {
                        "name": "cart_items",
                        "in": "query",
                        "description": "danh s\u00e1ch \u0111\u01a1n h\u00e0ng",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "raw_total",
                        "in": "query",
                        "description": "T\u1ed5ng ti\u1ec1n \u0111\u01a1n h\u00e0ng",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "total_qty",
                        "in": "query",
                        "description": "T\u1ed5ng s\u1ed1 l\u01b0\u1ee3ng s\u1ea3n ph\u1ea9m",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
    },
...
}